move changeAction to language module
