+++ src/app/static/src/app/store/dataExploration/actions.ts
@@ -1,12 +1,12 @@
 import {LanguageActions} from "../language/language";
 import {ActionTree} from "vuex";
 import {DataExplorationState} from "./dataExploration";
-import {ChangeLanguageHelper} from "../../utils";
+import {ChangeLanguageAction} from "../language/actions";
 
 export const actions: ActionTree<DataExplorationState, DataExplorationState> & LanguageActions<DataExplorationState> = {
 
     async changeLanguage(context, payload) {
-        await ChangeLanguageHelper(context, payload)
+        await ChangeLanguageAction(context, payload)
     }
 
 }
